chore: fix clippy warnings properly
- Fix manual_memcpy in csc.rs using copy_from_slice
- Replace Default impls with derive for FeatureSelectorKind, Objective, EvalMetric
- Fix clone on Copy types (ScalarLeaf) using dereference
- Fix immediate deref in predictor.rs
- Fix &PathBuf -> &Path in test helpers
- Fix manual RangeInclusive::contains usage
- Fix unnecessary cast in benchmark
- Use array instead of vec! for test data
- Allow unused_imports in tests/common re-exports
- Keep type_complexity and eq_op allows where intentional

+++ benches/linear_training.rs
@@ -501,7 +501,7 @@ fn bench_multiclass_training(c: &mut Criterion) {
         let (features, mut labels) = generate_training_data(num_rows, num_features, 42);
         // Convert labels to class indices
         for label in &mut labels {
-            *label = (label.abs() * num_classes as f32) as f32 % num_classes as f32;
+            *label = (label.abs() * num_classes as f32) % num_classes as f32;
         }
 
         let row_matrix = RowMatrix::from_vec(features, num_rows, num_features);
